AI_features/backend_agent/pyproject.toml
[project]
name = "agent-backend"
version = "0.1.0"
description = "Minimal LiveKit Agent Backend"
requires-python = ">=3.9"
dependencies = [
    "livekit-agents[silero,simli,turn-detector]~=1.3",
    "livekit-plugins-assemblyai>=1.3.10",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv",
    "fastapi",
    "uvicorn",
]

[tool.setuptools.packages.find]
where = ["."]
